makes the consistency checker a package under infra/
