+++ infra/consistency/src/validators/offer.js
@@ -4,11 +4,12 @@
  *
  * TODO: Add support for other events. (e.g. OfferAccepted, OfferFinalized, etc)
  */
+const esmImport = require('esm')(module)
 const db = {
-  ...require('../../../models')
+  ...esmImport('@origin/discovery/src/models')
 }
 
-const log = require('../../logger')
+const { log } = require('../logger')
 const { assert, getListenerBlock, getPastEvents } = require('./utils')
 
 async function offerCreatedEvents(contract, fromBlock, toBlock) {
